refactor: 메인 ViewController 이름 변경
