+++ WeatherForecast/WeatherForecast/Controller/WeatherListViewController.swift
@@ -1,0 +1,166 @@
+//
+//  WeatherListViewController - ViewController.swift
+//  Created by yagom. 
+//  Copyright © yagom. All rights reserved.
+// 
+
+import UIKit
+import CoreLocation
+
+class WeatherListViewController: UIViewController {
+
+    // MARK: - Properties
+
+    private let collectionView = UICollectionView(
+        frame: .zero,
+        collectionViewLayout: UICollectionViewFlowLayout()
+    )
+
+    private let repository = OpenWeatherRepository(
+        deserializer: JSONDesirializer(),
+        service: NetworkService()
+    )
+
+    private let locationDataManager = LocationDataManager()
+    private let addressManager = AddressManager()
+
+    // MARK: - Lifecycle
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupViewBacground()
+        setupLocationDataManager()
+        setupCollectionView()
+    }
+
+    override func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
+        collectionView.frame = view.bounds
+    }
+
+    // MARK: - Private
+
+    private func setupLocationDataManager() {
+        locationDataManager.delegate = self
+
+        if !locationDataManager.isAuthorized {
+            locationDataManager.requestAuthorization()
+        }
+    }
+
+    private func fetchLocation() {
+        locationDataManager.requestLocation()
+    }
+
+    private func fetchWeather(coordinate: Coordinate) {
+        repository.fetchWeather(coordinate: coordinate) { result in
+            switch result {
+            case .success(let currentWeather):
+                print(currentWeather.weathers.first?.description ?? "")
+            case .failure(let error):
+                print(error.localizedDescription)
+            }
+        }
+    }
+
+    private func fetchForecast(coordinate: Coordinate) {
+        repository.fetchForecast(coordinate: coordinate) { result in
+            switch result {
+            case .success(let forecast):
+                print(forecast.forecastDatas.first?.weathers.first?.description ?? "")
+            case .failure(let error):
+                print(error.localizedDescription)
+            }
+        }
+    }
+
+    private func setupCollectionView() {
+        view.addSubview(collectionView)
+        collectionView.dataSource = self
+        collectionView.delegate = self
+        collectionView.backgroundColor = .clear
+        collectionView.register(WeatherCell.self, forCellWithReuseIdentifier: WeatherCell.identifier)
+        collectionView.register(WeatherViewHeader.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: WeatherViewHeader.identifier)
+    }
+
+    private func setupViewBacground() {
+        let backgroundIamgeView: UIImageView = .init(frame: view.frame)
+        backgroundIamgeView.image = UIImage(named: "WeatherBackgroundImage")
+        backgroundIamgeView.contentMode = .scaleAspectFill
+
+        view.addSubview(backgroundIamgeView)
+        view.sendSubviewToBack(backgroundIamgeView)
+
+    }
+    
+}
+
+// MARK: - LocationDataManagerDelegate
+
+extension WeatherListViewController: LocationDataManagerDelegate {
+
+    func locationDataManager(_ locationDataManager: LocationDataManager,
+                             didAuthorized isAuthorized: Bool) {
+        guard isAuthorized else { return }
+        fetchLocation()
+    }
+
+    func locationDataManager(_ locationDataManager: LocationDataManager,
+                             didUpdateLocation location: CLLocation) {
+        addressManager.fetchAddress(of: location) { placemark in
+            guard let placemark else { return }
+
+            print(placemark.country ?? "",
+                  placemark.administrativeArea ?? "",
+                  placemark.locality ?? "",
+                  placemark.name ?? "")
+        }
+
+        guard let coordinate = locationDataManager.currentCoordinate else { return }
+
+        fetchWeather(coordinate: coordinate)
+        fetchForecast(coordinate: coordinate)
+    }
+}
+
+// MARK: - UICollectionViewDataSource
+
+extension WeatherListViewController: UICollectionViewDataSource {
+
+    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+        guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: WeatherViewHeader.identifier, for: indexPath) as? WeatherViewHeader else { return UICollectionReusableView() }
+
+        return header
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
+
+        return CGSize(width: collectionView.frame.width, height: collectionView.frame.height/10)
+    }
+
+    func collectionView(_ collectionView: UICollectionView,
+                        numberOfItemsInSection section: Int) -> Int {
+        30
+    }
+
+    func collectionView(_ collectionView: UICollectionView,
+                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: WeatherCell.identifier, for: indexPath) as? WeatherCell else { return UICollectionViewCell() }
+
+        cell.configure(with: String(indexPath.row))
+        return cell
+    }
+
+}
+
+// MARK: - UICollectionViewDelegateFlowLayout
+
+extension WeatherListViewController: UICollectionViewDelegateFlowLayout {
+
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: view.frame.size.width, height: view.frame.size.height / 15)
+    }
+
+}
