+++ WeatherForecast/WeatherForecast/Resources/Assets.xcassets/WeatherBackgroundImage.imageset/Contents.json
@@ -1,11 +1,11 @@
 {
   "images" : [
     {
-      "filename" : "WeatherBackgroundImage.jpg",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
+      "filename" : "WeatherBackgroundImage.jpeg",
       "idiom" : "universal",
       "scale" : "2x"
     },
